# Reduce overhead in tokenize