--- a/dask/tests/test_base.py
+++ b/dask/tests/test_base.py
@@ -34,10 +34,11 @@ from dask.base import (
     unpack_collections,
     visualize,
 )
+from dask.core import validate_key
 from dask.delayed import Delayed, delayed
 from dask.diagnostics import Profiler
 from dask.highlevelgraph import HighLevelGraph
-from dask.utils import tmpdir, tmpfile
+from dask.utils import key_split, tmpdir, tmpfile
 from dask.utils_test import dec, import_or_none, inc
 
 da = import_or_none("dask.array")
@@ -1011,15 +1012,13 @@ def test_optimizations_ctd():
 
 
 def test_clone_key():
-    assert clone_key("inc-1-2-3", 123) == "inc-73db79fdf4518507ddc84796726d4844"
-    assert clone_key("x", 123) == "x-c4fb64ccca807af85082413d7ef01721"
-    assert clone_key("x", 456) == "x-d4b538b4d4cf68fca214077609feebae"
-    assert clone_key(("x", 1), 456) == ("x-d4b538b4d4cf68fca214077609feebae", 1)
-    assert clone_key(("sum-1-2-3", h1, 1), 123) == (
-        "sum-822e7622aa1262cef988b3033c32aa37",
-        h1,
-        1,
-    )
+    for key, seed in [("x", 123), (("x", 1), 456), (("sum-1-2-3", h1, 1), 123)]:
+        validate_key(clone_key(key, seed))
+        assert clone_key(key, seed) != key
+        assert clone_key(key, seed) == clone_key(key, seed)
+        assert clone_key(key, seed) != clone_key(key, seed + 1)
+        assert key_split(clone_key(key, seed)) == key_split(key)
+
     with pytest.raises(TypeError):
         clone_key(1, 2)