Builds on https://github.com/dask/dask/pull/11371

I noticed that tokenize was using some non-trivial overhead. That overhead primarily came from the contextlib contextmanager but also from the ContextVars themselves. I threw both away...

The bottom line is an overhead reduction by about 60-70%. Further speedup would be possible if we ditched md5 instead of stdlib hash (or another hash function). It's not the fastest and I guess we don't need a very good hash function here (but I don't know)

# main (py3.10)

```python
In [1]: from dask.base import tokenize
   ...: trivial_items = [ix for ix in range(10000)]

In [2]: %timeit tokenize(trivial_items)
1.95 ms ± 282 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [3]: %timeit [tokenize(it) for it in trivial_items]
43.5 ms ± 489 μs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

# This PR

```python
In [1]: from dask.base import tokenize
   ...: trivial_items = [ix for ix in range(10000)]

In [2]: %timeit tokenize(trivial_items)
1.02 ms ± 9.43 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

In [3]: %timeit [tokenize(it) for it in trivial_items]
14.4 ms ± 142 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

I personally find this logic now also simpler and easier to reason about but this may be subjective. There is also a change in how memoization works. It now only memoizes collections if they are indeed recursive / self references. This may be slower in some (rare) situations but I believe it also avoids problems I was running into over in https://github.com/dask/dask/pull/11320

if performance for these edge cases are a problem we could choose to cache results instead (memoization remembers that a value was seen and is storing a dummy value, a cache would use the properly tokenized result instead of a dummy token)